requirements.txt
langchain==0.0.285
python-dotenv
openai
python-pptx